error

+++ demo/src/main/java/com/example/demo/controllers/AdminController.java
@@ -7,7 +7,6 @@ import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.servlet.ModelAndView;
 
 import com.example.demo.models.LocalBusinessOwner;
@@ -40,7 +39,7 @@ public class AdminController {
     }
 
     @GetMapping("/editBusinessOwner/{id}")
-    public ModelAndView editBusinessOwnerForm(@PathVariable Long id) {
+    public ModelAndView editBusinessOwnerForm(@PathVariable int id) {
         LocalBusinessOwner businessOwner = localBusinessOwnerRepository.findById(id)
             .orElseThrow(() -> new IllegalArgumentException("Invalid business owner Id:" + id));
         ModelAndView mav = new ModelAndView("admin/editBusinessOwner");
